Add proccess shutdown logic. Add logger (need to fix).

+++ Framework/utils/ProcessUtil.cs
@@ -4,16 +4,31 @@ using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using NLog;
 
 namespace Framework.utils
 {
     public static class ProcessUtil
     {
+        private static NLog.Logger Logger = LoggerUtil.GetLogger();
         public static void CloseAllProcessesByName(string name)
         {
+            Logger.Debug($"Kill all proccess with name = {name}");
+
             foreach (var process in Process.GetProcesses())
             {
-                process.Kill();
+                try
+                {
+                    if (process.ProcessName.Equals(name))
+                    {
+                        process.Kill();
+                    }
+                }
+                catch (Exception e)
+                {
+                    Logger.Warn($"Proccess with name = {process.ProcessName}) denied access. Exception message = {e.Message}");
+                    continue;
+                }
             }
         }
     }
